backend/requirements.txt
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
Django==5.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.8
django-cors-headers==3.14.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
oauthlib==3.2.0
psycopg2-binary==2.9.7
PyJWT==2.3.0
pyrsistent==0.18.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.5.3
Twisted==22.1.0
typing_extensions==4.14.0
urllib3==1.26.5
zope.interface==5.4.0
spacy==3.6.1
PyMuPDF==1.22.5
